add support for nginx configuration
Supporting a new language consists of implementing several things:

1. parser
2. annotator
3. 'tagged' documentation of the language

nginxparse.py is a modified parser based on the 'nginxparse' package. It
produces an AST that is later traversed by the annotator. The parser should
include positions in the original input.

The annotator outputs a list of annotations for the given input, in this case
an nginx config file. An annotation selects a range in the input, associates
a group with it and applies decoration. A group is just a string which appears
both in an annotation and somewhere in the docs in a 'data-showdocs' attribute,
where the relevant description appears.

Nginx documentation is built from a modified source repo in
external/nginx/build.bash. Currently each nginx directive is annotated. This
was surprsingly easy. Sometimes building documentation and modifying the
sources to include the data-showdocs in the output isn't as easy. A more naive
approach just takes the final html and edit that to include the data-showdocs
attrs.